spec: the low-level rollback is private to the module that wraps it
